+++ api/common/data-sources/index.ts
@@ -1,6 +1,4 @@
 export { usersDataSource } from "./users-data-source";
 export { authDataSource } from "./auth-provider-data-source";
-export { eventsDataSource } from "./events-data-source";
-export { ordersDataSource } from "./orders-data-source";
 export { orderProductsDataSource } from "./order-product-data-source";
 export { productsDataSource } from "./products-data-source";
