+++ api/common/data-sources/users-data-source.ts
@@ -2,7 +2,7 @@ import * as fp from "lodash/fp";
 import * as moment from "moment";
 import { v4 as uuid } from "uuid";
 
-import { user } from "../interfaces";
+import { user } from "../../partner/interfaces";
 
 const users: user.IUserDao[] = [
   {
