+++ api/partner/repository/order.ts
@@ -4,13 +4,17 @@ import * as boom from "boom";
 import { error, response } from "../utils";
 import { IOrder } from "../../common/models/order";
 import { IEvent } from "../../common/models/event";
-import { ordersDataSource, eventsDataSource } from "../../common/data-sources";
+import { IOrdersDataSource } from "../../common/data-sources/orders-data-source";
+import { IEventsDataSource } from "api/common/data-sources/events-data-source";
 
 const isFinished = (event: IEvent): boolean => {
   return event.expirationDate < Date.now() || event.markedAsFinished;
 };
 
-export const OrdersRepository = () => {
+export const OrdersRepository = (
+  ordersDataSource: IOrdersDataSource,
+  eventsDataSource: IEventsDataSource
+) => {
   const getOrders = async (): Promise<IOrder[]> => {
     return Promise.resolve(ordersDataSource.find());
   };
